Refactored GVK to be reusable
